refactor: extract stateless CalendarRepository helpers
