+++ app/src/main/java/com/kgs/calendar/data/CalendarRepositoryHelpers.kt
@@ -1,5 +1,8 @@
 package com.kgs.calendar.data
 
+import com.kgs.calendar.data.local.entity.CollectionEntity
+import org.json.JSONArray
+import java.io.IOException
 import java.nio.charset.StandardCharsets
 import java.util.UUID
 
@@ -29,4 +32,49 @@ internal fun String.looksLikeHtmlResponse(contentType: String?): Boolean {
     val prefix = trimStart().take(256)
     return prefix.startsWith("<!DOCTYPE html", ignoreCase = true) ||
         prefix.startsWith("<html", ignoreCase = true)
+}
+
+internal fun String?.updateAttendeePartstat(attendeeEmails: List<String>, partstat: String): String? {
+    val normalizedPartstat = partstat.trim().uppercase()
+    if (normalizedPartstat !in setOf("ACCEPTED", "DECLINED", "TENTATIVE", "NEEDS-ACTION")) return null
+    val matches = attendeeEmails
+        .map { it.trim().lowercase() }
+        .filter { it.isNotBlank() }
+        .toSet()
+    if (matches.isEmpty()) return null
+    val attendees = runCatching { JSONArray(this.orEmpty()) }.getOrNull() ?: return null
+    var changed = false
+    repeat(attendees.length()) { index ->
+        val obj = attendees.optJSONObject(index) ?: return@repeat
+        val email = obj.optString("email").trim().lowercase()
+        if (email in matches) {
+            obj.put("partstat", normalizedPartstat)
+            changed = true
+        }
+    }
+    return attendees.takeIf { changed }?.toString()
+}
+
+internal fun newUid(): String = "${UUID.randomUUID()}@kgs-calendar"
+
+internal fun CollectionEntity.resolvedAutomaticColor(): Int =
+    automaticColor ?: sourceColor ?: color
+
+internal fun Throwable.isTransientReadOnlySyncFailure(): Boolean {
+    if (findCause<IOException>() != null) return true
+    val statusCode = message
+        ?.substringAfter("URL returned HTTP ", "")
+        ?.takeWhile(Char::isDigit)
+        ?.toIntOrNull()
+        ?: return false
+    return statusCode == 408 || statusCode == 425 || statusCode == 429 || statusCode >= 500
+}
+
+internal inline fun <reified T : Throwable> Throwable.findCause(): T? {
+    var current: Throwable? = this
+    while (current != null) {
+        if (current is T) return current
+        current = current.cause
+    }
+    return null
 }
